[SC-95] fix device name in packer template
device_name "/dev/xvda" is the correct setting for aws linux AMIs.
setting it to this value for ubuntu will result in creation of two
snapshots for the image.  We really only need one root volume,
the fix is to set it to "/dev/sda1", this will create the AMI
will only one associated snapshot volume.